version openapi.yml
This patch versions the file api/openapi.yml. Even though this file is
generated by Typespec, it was not fully compatible with the oapi-codegen
tool, regarding all the allOf construct is used in the yaml.

Typespec genrated this way (for example):

    Movie:
      type: object
      required:
        - director
        - genre
        - status
      properties:
        director:
          type: string
        genre:
          type: string
        status:
          $ref: '#/components/schemas/MovieStatus'
      allOf:
        - $ref: '#/components/schemas/Media'

However, for the fields to be generated correctly in the models created
by the code generator, it should be like that:

    Movie:
      type: object
      required:
        - director
        - genre
        - status
      allOf:
        - $ref: '#/components/schemas/Media'
        - properties:
            director:
              type: string
            genre:
              type: string
            status:
              $ref: '#/components/schemas/MovieStatus'